Staticcamera (#17)
* Work on staticcamera with a detour in charControl cleanup

* MAtch all bute two staticcamera

* Update ObjectTransform from DKR

* WIP func_800427E8_433E8

* Half attempt at lobbycamPosition

* Match the rest of staticcamera

+++ include/variables.h
@@ -19,7 +19,7 @@ extern s32 controlKeys;
 extern s32 controlXjoy;
 extern s32 controlYjoy;
 extern f32 controlchr_gravity;
-extern s32 D_800A2DB4_A39B4[];
+extern f32 D_800A2DB4_A39B4[];
 
 //objects.c
 extern s8 D_800F2C60_F3860;
